test: 예외 응답 처리
