+++ src/main/java/roomescape/controller/reservation/ReservationController.java
@@ -28,7 +28,6 @@ public class ReservationController {
 
     @PostMapping
     public ResponseEntity<ReservationResponse> reserve(@RequestBody CreateReservationRequest request) {
-        //TODO : 날짜 바인딩 예외 응답 처리하기
         var reservation = service.reserve(request.name(), request.date(), request.timeSlotId(), request.themeId());
         var response = ReservationResponse.from(reservation);
         return ResponseEntity.created(URI.create("/reservations/" + reservation.id())).body(response);
@@ -48,10 +47,5 @@ public class ReservationController {
             return ResponseEntity.noContent().build();
         }
         return ResponseEntity.notFound().build();
-    }
-
-    @ExceptionHandler(value = IllegalArgumentException.class)
-    public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.BAD_REQUEST);
     }
 }
